update: extract amazon logo into component
