+++ apps/engmanagement/src/pages/amazon-logo.tsx
@@ -1,0 +1,25 @@
+const AmazonLogo = () => {
+  return (
+    <svg
+      className="w-28"
+      xmlns="http://www.w3.org/2000/svg"
+      fill="none"
+      viewBox="0 0 119 24"
+    >
+      <path
+        fill="#F90"
+        d="M47.607 18.095c-4.46 3.287-10.923 5.04-16.488 5.04-7.803 0-14.828-2.886-20.143-7.686-.417-.377-.043-.892.458-.598 5.736 3.337 12.827 5.345 20.152 5.345 4.94 0 10.376-1.022 15.373-3.144.755-.32 1.386.495.648 1.043"
+      />
+      <path
+        fill="#F90"
+        d="M49.46 15.973c-.567-.728-3.767-.344-5.203-.173-.438.053-.505-.328-.11-.602 2.548-1.793 6.73-1.276 7.218-.674.487.604-.127 4.796-2.522 6.797-.368.308-.718.144-.555-.263.538-1.343 1.744-4.353 1.173-5.085"
+      />
+      <path
+        fill="#fff"
+        d="M44.357 2.535V.791c0-.263.2-.44.44-.44h7.807c.25 0 .451.18.451.44v1.494c-.003.25-.214.578-.588 1.095l-4.045 5.776c1.503-.037 3.09.187 4.453.955.307.174.39.428.414.678v1.86c0 .255-.28.552-.575.398-2.401-1.26-5.591-1.396-8.247.013-.27.148-.554-.146-.554-.4v-1.767c0-.284.003-.769.287-1.2l4.686-6.72h-4.078c-.251 0-.451-.177-.451-.438ZM15.88 13.415h-2.375a.449.449 0 0 1-.424-.404V.82c0-.243.204-.437.458-.437h2.214c.23.01.414.187.431.407v1.594h.044C16.805.845 17.89.127 19.354.127c1.486 0 2.415.718 3.083 2.258.575-1.54 1.88-2.258 3.28-2.258.996 0 2.085.41 2.75 1.332.751 1.026.597 2.516.597 3.822l-.003 7.692a.449.449 0 0 1-.457.441h-2.372a.452.452 0 0 1-.428-.44v-6.46c0-.515.047-1.798-.066-2.285-.177-.819-.709-1.05-1.397-1.05-.574 0-1.176.385-1.42 1-.243.614-.22 1.643-.22 2.334v6.46a.449.449 0 0 1-.457.442h-2.372a.449.449 0 0 1-.427-.441l-.004-6.46c0-1.36.224-3.361-1.463-3.361-1.707 0-1.64 1.95-1.64 3.36v6.46a.45.45 0 0 1-.458.442ZM59.776.127C63.3.127 65.207 3.153 65.207 7c0 3.718-2.108 6.668-5.431 6.668-3.46 0-5.345-3.027-5.345-6.798 0-3.795 1.908-6.744 5.345-6.744Zm.02 2.488c-1.75 0-1.86 2.385-1.86 3.872 0 1.49-.024 4.67 1.84 4.67 1.84 0 1.927-2.566 1.927-4.13 0-1.028-.044-2.257-.354-3.233-.267-.848-.798-1.179-1.553-1.179Zm9.98 10.8h-2.364a.451.451 0 0 1-.428-.441L66.981.782a.45.45 0 0 1 .458-.398h2.2c.208.01.378.15.425.34v1.864h.044C70.772.922 71.704.127 73.344.127c1.066 0 2.105.384 2.773 1.436.621.975.621 2.615.621 3.795v7.672a.456.456 0 0 1-.458.385H73.9a.45.45 0 0 1-.421-.385V6.41c0-1.333.154-3.284-1.486-3.284-.578 0-1.11.388-1.373.976-.334.744-.378 1.486-.378 2.308v6.564a.457.457 0 0 1-.464.44Zm8.449-.97c0-.614.524-1.108 1.169-1.108.644 0 1.169.494 1.169 1.109 0 .611-.525 1.112-1.17 1.112-.644 0-1.168-.5-1.168-1.112Zm25.163.973a.446.446 0 0 1-.441-.44V.787a.44.44 0 0 1 .424-.4h.872c.24 0 .434.173.454.4v1.777c.621-1.41 1.777-2.512 3.237-2.512h.177c1.55 0 2.682 1.142 3.103 2.803.658-1.64 1.894-2.803 3.524-2.803h.18c1.153 0 2.262.742 2.836 1.87.555 1.08.535 2.516.535 3.722l-.004 7.332a.445.445 0 0 1-.424.44h-1.042a.456.456 0 0 1-.441-.38V5.645c0-.872.043-1.797-.311-2.592-.36-.812-1.052-1.32-1.793-1.356-.829.04-1.59.644-2.085 1.46-.641 1.052-.618 2-.618 3.23v6.657a.457.457 0 0 1-.424.374h-1.036a.45.45 0 0 1-.444-.44l-.006-7.82c0-.719-.044-1.564-.375-2.209-.384-.735-1.072-1.216-1.793-1.253-.749.044-1.504.615-1.951 1.28-.578.848-.688 1.897-.688 2.976v7.025a.452.452 0 0 1-.428.44h-1.038Zm-7.156.254c-3.37 0-4.88-3.437-4.88-6.875 0-3.614 1.774-6.744 5.168-6.744h.18c3.3 0 4.944 3.334 4.944 6.771 0 3.641-1.82 6.848-5.234 6.848h-.177Zm.248-1.667c1.109-.037 1.984-.725 2.502-1.87.464-1.03.554-2.209.554-3.338 0-1.229-.133-2.512-.711-3.564-.518-.918-1.407-1.503-2.349-1.536-1.049.036-1.987.748-2.458 1.844-.424.948-.555 2.208-.555 3.256 0 1.18.154 2.542.642 3.568.474.972 1.386 1.603 2.375 1.64ZM85.7 11.95c1.514-.046 2.308-1.259 2.636-2.829.067-.197.217-.347.438-.347l.998-.004c.238.01.455.19.435.411-.461 2.676-2.075 4.49-4.373 4.49h-.18c-3.347 0-4.79-3.36-4.79-6.771 0-3.384 1.463-6.848 4.81-6.848h.18c2.325 0 3.982 1.79 4.343 4.466 0 .2-.188.374-.408.398l-1.045-.014c-.22-.03-.364-.217-.398-.427-.25-1.493-1.109-2.683-2.539-2.73-2.275.074-2.923 2.873-2.923 5.028 0 2.074.541 5.104 2.816 5.177Zm-43.338-.771c-.434-.601-.895-1.09-.895-2.201v-3.7c0-1.567.11-3.007-1.045-4.086C39.51.317 38 .01 36.844.01c-2.258 0-4.783.841-5.311 3.634-.057.297.16.454.354.498l2.301.247c.218-.01.375-.22.415-.434.197-.962 1.002-1.427 1.907-1.427.491 0 1.046.18 1.333.618.334.488.29 1.156.29 1.72v.308c-1.376.157-3.176.257-4.466.825-1.49.641-2.532 1.954-2.532 3.882 0 2.468 1.553 3.7 3.555 3.7 1.686 0 2.612-.397 3.915-1.726.43.624.574.928 1.362 1.583a.496.496 0 0 0 .562-.057l.003.007c.474-.421 1.336-1.172 1.82-1.577.194-.16.16-.417.01-.631ZM37.69 10.11c-.377.668-.978 1.079-1.643 1.079-.912 0-1.447-.695-1.447-1.72 0-2.025 1.814-2.392 3.535-2.392v.514c0 .926.023 1.697-.445 2.52Zm-26.463 1.07c-.434-.601-.895-1.09-.895-2.201V5.278c0-1.567.11-3.007-1.046-4.086C8.374.317 6.864.01 5.71.01 3.45.01.929.85.398 3.644c-.054.297.16.454.353.498l2.305.247c.214-.01.371-.22.411-.434.197-.962 1.006-1.427 1.911-1.427.488 0 1.042.18 1.333.618.33.488.287 1.156.287 1.72v.308c-1.376.157-3.177.257-4.466.825C1.046 6.64 0 7.953 0 9.881c0 2.468 1.557 3.7 3.554 3.7 1.69 0 2.612-.397 3.915-1.726.434.624.575.928 1.363 1.583a.496.496 0 0 0 .561-.057l.007.007a99.556 99.556 0 0 1 1.82-1.577c.194-.16.16-.417.007-.631Zm-4.673-1.07c-.378.668-.976 1.079-1.644 1.079-.912 0-1.443-.695-1.443-1.72 0-2.025 1.814-2.392 3.531-2.392v.514c0 .926.023 1.697-.444 2.52Z"
+      />
+    </svg>
+  )
+}
+
+export default AmazonLogo
